requirements.txt
Django==3.1.1
djangorestframework==3.12.4
requests